fix(tooling): allow concurrent worktree validation
